Retrieve publishers from instance methods
