+++ Sources/Player/PlayerItemPublishers.swift
@@ -1,0 +1,21 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import AVFoundation
+import Combine
+
+extension AVPlayerItem {
+    func itemStatePublisher() -> AnyPublisher<ItemState, Never> {
+        Publishers.Merge(
+            publisher(for: \.status)
+            // TODO: Weakify? KVO on error? Write UT to check for AVPlayerItem deallocation
+                .map { ItemState.itemState(for: $0, error: self.error) },
+            NotificationCenter.default.weakPublisher(for: .AVPlayerItemDidPlayToEndTime, object: self)
+                .map { _ in .ended }
+        )
+        .eraseToAnyPublisher()
+    }
+}
