--- Sources/Player/Publishers.swift
@@ -1,12 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import Foundation
-
-// TODO: Other ideas for publishers
-//        - Boundary time publisher
-//        - Asset property publisher (can register to one or several properties and have them published; implemented
-//          internally with modern async API). Might be useful as signal for the Pulse publisher.
