+++ Sources/Player/StreamType.swift
@@ -4,7 +4,7 @@
 //  License information is available from the LICENSE file.
 //
 
-import AVFoundation
+import Foundation
 
 /// Stream types.
 public enum StreamType {
