+++ Tests/PlayerTests/PulseTests.swift
@@ -6,8 +6,7 @@
 
 @testable import Player
 
-import AVFoundation
-import Circumspect
+import CoreMedia
 import Nimble
 import XCTest
 
@@ -98,170 +97,6 @@ final class PulseTimeTests: XCTestCase {
     func testTimeForProgressWithEmptyTimeRange() {
         let pulse = Pulse(time: .zero, timeRange: .zero, itemDuration: .indefinite)!
         expect(pulse.time(forProgress: 2)).to(beNil())
-    }
-}
-
-final class SingleItemPulsePublisherTests: XCTestCase {
-    func testOnDemandPulse() throws {
-        let item = AVPlayerItem(url: TestStreams.shortOnDemandUrl)
-        let player = AVPlayer(playerItem: item)
-        try expectPublished(
-            values: [
-                Pulse(
-                    time: .zero,
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: CMTime(value: 1, timescale: 1),
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                )
-            ],
-            from: Pulse.publisher(for: player, interval: CMTime(value: 1, timescale: 1), queue: .main),
-            during: 4
-        ) {
-            player.play()
-        }
-    }
-
-    func testLivePulse() throws {
-        let item = AVPlayerItem(url: TestStreams.liveUrl)
-        let player = AVPlayer(playerItem: item)
-        try expectPublished(
-            values: [
-                Pulse(time: .zero, timeRange: .zero, itemDuration: .indefinite),
-                Pulse(time: CMTime(value: 1, timescale: 1), timeRange: .zero, itemDuration: .indefinite),
-                Pulse(time: CMTime(value: 2, timescale: 1), timeRange: .zero, itemDuration: .indefinite)
-            ],
-            from: Pulse.publisher(for: player, interval: CMTime(value: 1, timescale: 1), queue: .main)
-        ) {
-            player.play()
-        }
-    }
-
-    func testFailedPlaybackPulse() throws {
-        let item = AVPlayerItem(url: TestStreams.unavailableUrl)
-        let player = AVPlayer(playerItem: item)
-        try expectPublished(
-            values: [],
-            from: Pulse.publisher(for: player, interval: CMTime(value: 1, timescale: 1), queue: .main),
-            during: 2
-        ) {
-            player.play()
-        }
-    }
-
-    func testNonPlayingPulse() throws {
-        let player = AVPlayer()
-        try expectPublished(
-            values: [],
-            from: Pulse.publisher(for: player, interval: CMTime(value: 1, timescale: 1), queue: .main),
-            during: 2
-        ) {
-            player.play()
-        }
-    }
-}
-
-final class MultipleItemPulsePublisherTests: XCTestCase {
-    func testChainedShortItems() throws {
-        let item1 = AVPlayerItem(url: TestStreams.shortOnDemandUrl)
-        let item2 = AVPlayerItem(url: TestStreams.shortOnDemandUrl)
-        let player = AVQueuePlayer(items: [item1, item2])
-        try expectPublished(
-            values: [
-                Pulse(
-                    time: .zero,
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: CMTime(value: 1, timescale: 1),
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: .zero,
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: CMTime(value: 1, timescale: 1),
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                )
-            ],
-            from: Pulse.publisher(for: player, interval: CMTime(value: 1, timescale: 1), queue: .main),
-            during: 4
-        ) {
-            player.play()
-        }
-    }
-
-    func testChainedItemsWithFailure() throws {
-        let item1 = AVPlayerItem(url: TestStreams.shortOnDemandUrl)
-        let item2 = AVPlayerItem(url: TestStreams.unavailableUrl)
-        let item3 = AVPlayerItem(url: TestStreams.shortOnDemandUrl)
-        let player = AVQueuePlayer(items: [item1, item2, item3])
-        try expectPublished(
-            values: [
-                Pulse(
-                    time: .zero,
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: CMTime(value: 1, timescale: 1),
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: .zero,
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                ),
-                Pulse(
-                    time: CMTime(value: 1, timescale: 1),
-                    timeRange: CMTimeRange(
-                        start: .zero,
-                        duration: CMTime(value: 1, timescale: 1)
-                    ),
-                    itemDuration: CMTime(value: 1, timescale: 1)
-                )
-            ],
-            from: Pulse.publisher(for: player, interval: CMTime(value: 1, timescale: 1), queue: .main),
-            during: 4
-        ) {
-            player.play()
-        }
     }
 }
